Restructure the connection and client classes
- Splits the client into two classes (client + connection) to better focus on
  the purpose of the class.
- Introduce a `configurable` class which is used to configure the client.
- Update specs to utilise `test_api_token` and `test_api_username` methods.
