+++ spec/cassettes/client/not_found.json
@@ -1,1 +1,1 @@
-{"http_interactions":[{"request":{"method":"get","uri":"https://api.envato.com/v1/non_existent_path","body":{"encoding":"US-ASCII","string":""},"headers":{"User-Agent":["Envato SDK (0.0.1)"],"Authorization":["Bearer <ENVATO_API_TOKEN>"],"Accept-Encoding":["gzip;q=1.0,deflate;q=0.6,identity;q=0.3"],"Accept":["*/*"]}},"response":{"status":{"code":404,"message":"Not Found"},"headers":{"Access-Control-Allow-Origin":["*"],"Cache-Control":["max-age=0, private, no-cache, no-store, must-revalidate"],"Content-Type":["text/html; charset=utf-8"],"Etag":["W/\"9-70b5a349\""],"Expires":["0"],"Server":["nginx/1.6.2"],"Vary":["*"],"X-Frame-Options":["Deny"],"X-Powered-By":["Express"],"Content-Length":["9"],"Date":["Thu, 15 Oct 2015 06:34:44 GMT"],"Connection":["keep-alive"]},"body":{"encoding":"UTF-8","string":"Not found"},"http_version":null},"recorded_at":"Thu, 15 Oct 2015 06:34:44 GMT"}],"recorded_with":"VCR 2.7.0"}
+{"http_interactions":[{"request":{"method":"get","uri":"https://api.envato.com/v1/non_existent_path","body":{"encoding":"US-ASCII","string":""},"headers":{"User-Agent":["Envato SDK (0.0.1)"],"Authorization":["Bearer <ENVATO_API_TOKEN>"],"Accept-Encoding":["gzip;q=1.0,deflate;q=0.6,identity;q=0.3"],"Accept":["*/*"]}},"response":{"status":{"code":404,"message":"Not Found"},"headers":{"Access-Control-Allow-Origin":["*"],"Cache-Control":["max-age=0, private, no-cache, no-store, must-revalidate"],"Content-Type":["text/html; charset=utf-8"],"Etag":["W/\"9-70b5a349\""],"Expires":["0"],"Server":["nginx/1.6.2"],"Vary":["*"],"X-Frame-Options":["Deny"],"X-Powered-By":["Express"],"Content-Length":["9"],"Date":["Sat, 17 Oct 2015 06:51:42 GMT"],"Connection":["keep-alive"]},"body":{"encoding":"UTF-8","string":"Not found"},"http_version":null},"recorded_at":"Sat, 17 Oct 2015 06:51:42 GMT"}],"recorded_with":"VCR 2.9.3"}
