+++ spec/cassettes/client/valid_response.json
@@ -1,1 +1,1 @@
-{"http_interactions":[{"request":{"method":"get","uri":"https://api.envato.com/v1/market/total-items.json","body":{"encoding":"US-ASCII","string":""},"headers":{"User-Agent":["Envato SDK (0.0.1)"],"Authorization":["Bearer <ENVATO_API_TOKEN>"],"Accept-Encoding":["gzip;q=1.0,deflate;q=0.6,identity;q=0.3"],"Accept":["*/*"]}},"response":{"status":{"code":200,"message":"OK"},"headers":{"Access-Control-Allow-Origin":["*"],"Cache-Control":["max-age=0, private, no-cache, no-store, must-revalidate"],"Content-Type":["application/json"],"Etag":["W/\"29-4bcf8989\""],"Expires":["0"],"Last-Modified":["Thu, 15 Oct 2015 06:32:57 GMT"],"Server":["nginx/1.6.2"],"Vary":["*"],"X-Frame-Options":["Deny"],"X-Powered-By":["Express"],"X-Ratelimit-Remaining":["99"],"X-Ratelimit-Total":["100"],"Content-Length":["41"],"Date":["Thu, 15 Oct 2015 06:34:43 GMT"],"Connection":["keep-alive"]},"body":{"encoding":"UTF-8","string":"{\"total-items\":{\"total_items\":\"8676143\"}}"},"http_version":null},"recorded_at":"Thu, 15 Oct 2015 06:34:43 GMT"}],"recorded_with":"VCR 2.7.0"}
+{"http_interactions":[{"request":{"method":"get","uri":"https://api.envato.com/v1/market/total-items.json","body":{"encoding":"US-ASCII","string":""},"headers":{"User-Agent":["Envato SDK (0.0.1)"],"Authorization":["Bearer <ENVATO_API_TOKEN>"],"Accept-Encoding":["gzip;q=1.0,deflate;q=0.6,identity;q=0.3"],"Accept":["*/*"]}},"response":{"status":{"code":200,"message":"OK"},"headers":{"Access-Control-Allow-Origin":["*"],"Cache-Control":["max-age=0, private, no-cache, no-store, must-revalidate"],"Content-Type":["application/json"],"Etag":["W/\"29-f19417bd\""],"Expires":["0"],"Last-Modified":["Sat, 17 Oct 2015 06:41:53 GMT"],"Server":["nginx/1.6.2"],"Vary":["*"],"X-Frame-Options":["Deny"],"X-Powered-By":["Express"],"X-Ratelimit-Remaining":["99"],"X-Ratelimit-Total":["100"],"Content-Length":["41"],"Date":["Sat, 17 Oct 2015 06:51:41 GMT"],"Connection":["keep-alive"]},"body":{"encoding":"UTF-8","string":"{\"total-items\":{\"total_items\":\"8696472\"}}"},"http_version":null},"recorded_at":"Sat, 17 Oct 2015 06:51:41 GMT"}],"recorded_with":"VCR 2.9.3"}
